requirements.txt
anthropic>=0.49.0
pydantic>=2.0.0
numpy>=1.26.0
scipy>=1.12.0
pandas>=2.2.0
python-dotenv>=1.0.0
rich>=13.7.0
typer>=0.12.0
structlog>=24.1.0
asyncio-throttle>=1.0.2
